+++ source/deprecated/omni.isaac.camera_inspector/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.0.4] - 2025-03-26
+### Changed
+- Cleanup and standardize extension.toml, update code formatting for all code
+
 ## [1.0.3] - 2025-03-20
 ### Changed
 - Fix issues with duplicate extension startup from extra imports