+++ source/deprecated/omni.isaac.sensor/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [13.0.5] - 2025-03-26
+### Changed
+- Cleanup and standardize extension.toml, update code formatting for all code
+
 ## [13.0.4] - 2025-03-25
 ### Changed
 - Add import tests for deprecated extensions