+++ source/extensions/isaacsim.core.cloner/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.3.7] - 2025-03-26
+### Changed
+- Cleanup and standardize extension.toml, update code formatting for all code
+
 ## [1.3.6] - 2025-03-11
 ### Changed
 - Switch asset root for tests to internal nucleus