+++ source/extensions/isaacsim.gui.property/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [1.1.1] - 2025-03-26
+### Changed
+- Cleanup and standardize extension.toml, update code formatting for all code
 
 ## [1.1.0] - 2025-03-06
 ### Added