+++ source/extensions/isaacsim.replicator.behavior.ui/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [1.0.2] - 2025-03-26
+### Changed
+- Cleanup and standardize extension.toml, update code formatting for all code
 
 ## [1.0.1] - 2025-01-21
 ### Changed