+++ source/extensions/isaacsim.replicator.behavior/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.0.11] - 2025-03-26
+### Changed
+- Cleanup and standardize extension.toml, update code formatting for all code
+
 ## [1.0.10] - 2025-03-24
 ### Changed
 - Migrate to Events 2.0