+++ source/extensions/isaacsim.robot_setup.xrdf_editor/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [3.1.6] - 2025-03-26
+### Changed
+- Cleanup and standardize extension.toml, update code formatting for all code
 
 ## [3.1.5] - 2025-01-27
 ### Changed