+++ source/extensions/isaacsim.sensors.physics.examples/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [0.1.6] - 2025-03-26
+### Changed
+- Cleanup and standardize extension.toml, update code formatting for all code
+
 ## [0.1.5] - 2025-03-11
 ### Changed
 - Switch asset root for tests to internal nucleus