+++ source/extensions/isaacsim.xr.openxr/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [0.2.9] - 2025-03-26
+### Changed
+- Cleanup and standardize extension.toml, update code formatting for all code
+
 ## [0.2.8] - 2025-03-05
 ### Changed
 - Update extension codebase to adhere to isaac sim extension structure and file naming  guidelines